backend/requirements.txt
Django==5.2
Pillow==11.2.1
djangorestframework==3.16.0
drf-yasg==1.21.10
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-decouple==3.8
django-jazzmin==3.0.1
qrcode==8.2
gunicorn==23.0.0
whitenoise==6.9.0